fix(mobile): keep thread and terminal back navigation, and settings reachable hostless
Gating those two screens through a layout made each the root of its own stack,
which strips the back button; a Slot layout loses the header instead. The gate
moves into the screens so both stay screens of the host stack.

Settings mounts at the root route as well as the tab: nothing in it is
host-scoped, and the root mount is what keeps Manage hosts reachable with no host
registered. Its flow now guards on a row, since a host screen carries a Settings
tab and the title alone no longer identifies the screen.
