+++ apps/mobile/src/app/settings.tsx
@@ -1,0 +1,6 @@
+import { SettingsScreen } from '@mobile/components/settings/settings-screen';
+
+/** Root mount: reachable by deep link and with no host registered at all. */
+export default function SettingsRoute(): React.ReactNode {
+  return <SettingsScreen />;
+}
